+++ template/frontend/src/app/charts/base-chart/ChartConfig.ts
@@ -1,0 +1,7 @@
+import {ChartCallbackFunction, Options} from "highcharts";
+
+export type ChartConfig = {
+  id: string;
+  options: Options;
+  callback?: ChartCallbackFunction;
+}