+++ template/frontend/src/app/charts/scores-chart/ScoresChartConfig.ts
@@ -1,0 +1,4 @@
+export class ScoresChartConfig {
+  instance_name: string = "";
+  reference_value?: number;
+}