+++ template/frontend/src/app/model/Solution.ts
@@ -1,0 +1,12 @@
+export class Solution {
+  lastMoves: Move[] = [];
+  // DRFP specific
+  solutionData: any[] = [];
+  rowSizes: number[] = [];
+}
+
+
+
+export class Move {
+
+}